Restore metrics collector in management plugin
Summary:
The base image disable explicitly fot the management plugin
the metrics collector and associated UI to avoid redundancy
with Prometheus metrics.

Metrics in the UI for queues is something useful to debug
in an unified place, even with dashboards on the side.

For a discussion on metrics collection, see also T1623 and T1633.

Test Plan: Deploy for white-rabbit and orange-rabbit

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3158